Modif

+++ codeC/basics.h
@@ -18,21 +18,21 @@
 // Structures
 typedef struct Station {
     uint32_t id;
-    long consumption_sum;
-    long capacity;
+    unsigned long consumption_sum;
+    unsigned long capacity;
     char type[MAX_STATION_TYPE]; // HVA, HVB or LV
 } Station;
 
 typedef struct {
-    uint32_t id;
-    long consumption;
+    long id;
+    unsigned long consumption;
 } Consumer;
 
 typedef struct {
     Station station;
     struct AVL *left;
     struct AVL *right;
-    int balance;
+    int balance; // on verra
 } AVL;
 
 typedef AVL* pAVL;
